Changed logic in Parser _readHeaderLine
When the header line is present and the settings ask for it  to be processed, two different options are possible:

a) The schema has been populated.  In this case, build a new  schema where the order matches the *actual* order in which the given CSV file offers its columns; there cases the consumer of the csv file knows about the columns but not necessarily the order in which they are defined.  A further check can be done in this case, by not permitting new columns that are not defined in the schema, by adding a new flag to the schema (say, strict) and reporting an error when a new column is found in the header.

b) The schema has not been populated.  In this case, build a default schema based on the columns found in the header.