Add trailing comma feature
Currently the behavior allows there to be a single extraneous data column
that is empty. This is normally harmless, however, external tools that
expect the same number of data columns as header columns will complain
when this extraneous column is encountered.

The fix is to make this feature explicit and allow for it to be disabled.

This patch retains backwards compatibility by enabling the feature by
default.